LPD-89578 Add custom plugin example to the CKEditor 5 by Client Extension
