+++ modules/apps/frontend-editor/frontend-editor-ckeditor5-sample-web/src/main/resources/META-INF/resources/js/CKEditor5ReactClassicEditor.tsx
@@ -13,8 +13,6 @@ import {
 } from 'frontend-editor-ckeditor-web';
 import React, {useRef, useState} from 'react';
 
-import Timestamp from './Timestamp';
-
 const CKEditor5ReactClassicEditor = ({
 	editorConfig,
 }: {
@@ -26,7 +24,7 @@ const CKEditor5ReactClassicEditor = ({
 
 	const config: LiferayEditorConfig = {
 		...editorConfig,
-		extraPlugins: [Bookmark, Timestamp],
+		extraPlugins: [Bookmark],
 		initialData:
 			'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nunc id cursus metus aliquam eleifend mi in nulla. Quam adipiscing vitae proin sagittis nisl rhoncus. Suspendisse faucibus interdum posuere lorem. Nullam ac tortor vitae purus faucibus ornare. Ac felis donec et odio pellentesque diam. Nulla at volutpat diam ut. Posuere urna nec tincidunt praesent semper feugiat nibh. Gravida quis blandit turpis cursus. Proin libero nunc consequat interdum varius. Sollicitudin ac orci phasellus egestas tellus rutrum tellus pellentesque. Neque volutpat ac tincidunt vitae semper quis lectus nulla at. Odio euismod lacinia at quis risus sed vulputate odio ut. Augue lacus viverra vitae congue eu consequat ac. Elementum sagittis vitae et leo duis ut diam. Diam quis enim lobortis scelerisque fermentum dui faucibus. <p><a href="/home">Link to home page</a></p>',
 		removePlugins: [Underline],
@@ -40,7 +38,6 @@ const CKEditor5ReactClassicEditor = ({
 				'underline',
 				'|',
 				'bookmark',
-				'timestamp',
 				'|',
 				'headlessImageSelector',
 				'headlessVideoSelector',
