+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-3/client-extension.dev.yaml
@@ -1,0 +1,2 @@
+liferay-sample-editor-config-contributor-3:
+    baseURL: http://localhost:3002/build
