+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-3/client-extension.yaml
@@ -1,0 +1,9 @@
+assemble:
+    -   from: build
+        into: static
+liferay-sample-editor-config-contributor-3:
+    editorConfigKeys:
+        -   sampleReactCKEditor5ClassicEditor
+    name: Liferay Sample Editor Config Contributor CKEditor 5 Timestamp
+    type: editorConfigContributor
+    url: index.js
