+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-3/tsconfig.json
@@ -1,0 +1,21 @@
+{
+	"compilerOptions": {
+		"esModuleInterop": true,
+		"jsx": "react",
+		"lib": [
+			"DOM",
+			"ES2020"
+		],
+		"module": "ES2020",
+		"moduleResolution": "node",
+		"outDir": "./build",
+		"resolveJsonModule": true,
+		"sourceMap": false,
+		"target": "ES2020",
+		"types": [
+		]
+	},
+	"include": [
+		"./src/**/*"
+	]
+}
