+++ workspaces/liferay-sample-workspace/package.json
@@ -19,6 +19,7 @@
 			"client-extensions/liferay-sample-custom-element-5",
 			"client-extensions/liferay-sample-editor-config-contributor-1",
 			"client-extensions/liferay-sample-editor-config-contributor-2",
+			"client-extensions/liferay-sample-editor-config-contributor-3",
 			"client-extensions/liferay-sample-etc-node",
 			"client-extensions/liferay-sample-fds-cell-renderer",
 			"client-extensions/liferay-sample-fds-filter",
